[remove] Robokoz font
The 'Public Domain' font told by certain fonts websites is not certified ; other website provide the same font with a README from the author stating : 
"No payment is required for the use of this font. It's free!
Commercial or personal use? It doesn't matter, it's free."
(ref:  http://dafont.me/view_file/175/UmVhZF9NZS50eHQ=/read_me.txt )
This statements unfortunately doesn't explicit 'public domain or CC-0 , and so, it's difficult to now if derivations are accepted.
(Note : no problem with keeping the file in Git tree , file is unchanged, and correctly attributed all along.)